COSM: move is_relevant functions into utils to avoid circular import references
